+++ pages/portfolios/[id].js
@@ -3,25 +3,26 @@
 //Date: 08/06/2020
 //Description: This is the dynamically created page for each project in the porfolio.
 import React,{Component} from 'react'
-import BaseLayout from '../components/layouts/baselayout'
-import BasePage from '../components/BasePage'
+import BaseLayout from '../../components/layouts/baselayout'
+import BasePage from '../../components/BasePage'
 import {withRouter} from 'next/router'
 import axios from 'axios'
 
 //functional class based component
 
-const Portfolio=()=>{
-    const {portfolio} = props;
+const Portfolio=({portfolio})=>{
+
 
         return(
             <BaseLayout>
                 <BasePage>
-
-                </BasePage>
                 <h1>I am a project</h1>
                 <h1>{portfolio.title}</h1>
                 <p>BODY:{portfolio.body}</p>
                 <p>ID:{portfolio.id}</p>
+
+                </BasePage>
+
             </BaseLayout>
         )
 }
