+++ server.js
@@ -5,11 +5,11 @@
 
 const express = require('express');
 const next = require('next')
-const routes = require('./routes')
+
 
 const dev = process.env.NODE_ENV !== 'production'
 const app = next({dev})
-const handle = routes.getRequestHandler(app);
+const handle = app.getRequestHandler();
